refactor: ProductAll 컴포넌트의 추상화 고려 리팩토링
